use data values
